requirements.txt
tensorflow== 2.15.0
flask>=2.0.0,<3.0.0
fastapi==0.88.0
uvicorn==0.20.0
numpy==1.26.3
Pillow==10.2.0